Rewrite Linux popen to fiddle with sockets directly. (Ref #6606)
EM.popen does a bunch of unpleasant things like not keep track of the processes at all; this reimplementation has two attractive features.
- less magic
- support for multi-argument exec

One thing it doesn't do yet is use SIGCHLD to tell it to reap a child, or exploit the kqueue functionality.  The kqueue functionality is a WONTFIX because it's broken (no surprise).  SIGCHLD is hard to figure out how it would work in this context but I am open to adding that.